% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_raster.R
\name{as_raster.bivden}
\alias{as_raster.bivden}
\title{Converters for sparr objects to rasters}
\usage{
as_raster.bivden(x, crs, crop = NULL)
}
\arguments{
\item{x}{An object of class \code{bivden}}

\item{crs}{Coordinate reference system, something suitable
as input to \code{st_crs}}

\item{crop}{\code{sf} object to which to crop the raster}
}
\value{
A raster
}
\description{
Converters for sparr objects to rasters
}
