% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nw_europe}
\alias{nw_europe}
\title{Northwestern Europe country borders.}
\format{
An object of class \code{sfc_MULTIPOLYGON}.
}
\source{
\url{https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/10m/cultural/ne_10m_admin_1_states.zip}
}
\usage{
nw_europe
}
\description{
A \code{MULTIPOLYGON} of Northwestern Europe.
}
\keyword{datasets}
