% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p3k14c_data}
\alias{p3k14c_data}
\title{The scrubbed and fuzzed p3k14c dataset}
\format{
An object of class \code{tibble}.
}
\source{
\url{https://core.tdar.org/dataset/459163/p3k14c}
}
\usage{
p3k14c_data
}
\description{
The dates in this dataset were gathered from 40 pre-existing datasets,
verified, and cleaned according to a standardized sample selection criteria.
Dates for the United States and Canada are aggregated to county level,
while locational information for dates from one dataset
(Guedes and Bocinsky 2018) have been truncated to 2 decimal points.
All dates are uncalibrated.
}
\keyword{datasets}
