% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{A custom theme for ggplot maps.}
\usage{
theme_map(...)
}
\arguments{
\item{...}{Parameters passed on to \code{theme}}
}
\description{
A custom theme for ggplot maps.
}
