% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window.ppp}
\alias{window.ppp}
\title{Tidy function to window a Point Pattern}
\usage{
\method{window}{ppp}(x, w)
}
\arguments{
\item{x}{An object of class \code{ppp}}

\item{w}{An object of class \code{owin}}
}
\value{
An object of class \code{ppp} windowed to \code{w}
}
\description{
Tidy function to window a Point Pattern
}
