% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world}
\alias{world}
\title{World land area.}
\format{
An object of class \code{sfc_MULTIPOLYGON}.
}
\source{
\url{https://www.naturalearthdata.com/http//www.naturalearthdata.com/download/110m/physical/ne_110m_land.zip}
}
\usage{
world
}
\description{
A \code{MULTIPOLYGON} of global land area.
}
\keyword{datasets}
